troubleshoot status: extract --report rendering into helper module
Move the ~340-line report renderer and its 12 helpers out of
custom.py into a dedicated module (_troubleshoot_status_report.py)
with a single public entry point, render_report(payload). custom.py
now lazily imports it only on the --report path.

No behavioral change; keeps custom.py review-friendly.

Copilot-Session: da04b3ec-d5e4-4ab8-aeea-58dd7010ac84
